Using Modal service to confirm deletion of an activity

+++ client/app/admin/blog-mgmt/blog-mgmt.controller.js
@@ -1,7 +1,7 @@
 'use strict';
 
 angular.module('gripirBlogApp')
-  .controller('BlogMgmtCtrl', function ($scope, $http, socket) {
+  .controller('BlogMgmtCtrl', function ($scope, $http, Modal, socket) {
     $scope.activities = [];
 
     $http.get('/api/activities').success(function(activities) {
@@ -13,15 +13,14 @@ angular.module('gripirBlogApp')
       if($scope.newActivity === '') {
         return;
       }
+
       $http.post('/api/activities', { name: $scope.newActivity });
       $scope.newActivity = '';
     };
 
-    $scope.deleteActivity = function(activity) {
-
-      // TODO: Confirm dialog
+    $scope.deleteActivity = Modal.confirm.delete(function(activity) {
       $http.delete('/api/activities/' + activity._id);
-    };
+    });
 
     $scope.$on('$destroy', function () {
       socket.unsyncUpdates('activity');
